requirements.txt
alembic==1.8.1
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.7
Flask-SQLAlchemy==2.4.3
importlib-metadata==4.12.0
importlib-resources==5.9.0
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.2.1
MarkupSafe==1.1.1
mysqlclient==1.4.2.post1
PyMySQL==1.0.2
python-dotenv==0.20.0
six==1.15.0
SQLAlchemy==1.3.18
SQLAlchemy-Utils==0.29.8
Werkzeug==1.0.1
zipp==3.8.1